requirements.txt
fastapi
uvicorn
mysql-connector-python
pyyaml
Flask-SQLAlchemy
PyJWT
Werkzeug
